Refactor Grid stuff

+++ src/editor/Editor.h
@@ -46,7 +46,7 @@ public:
 	// Grid properties
 	static bool show_grid_;
 	static bool snap_to_grid_;
-	static float grid_size_;
+	static int grid_size_;
 
 private:
 	// List of all windows in the Editor
